Abhishek Commit (28-05-2024 10:54 Am)

+++ sbi-colending/src/main/java/com/sbicolending/model/createloanresponse/FinancialDataModel.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class FinancialDataModel {
 
-      private int annual_business_turnover;
+      private float annual_business_turnover;
 }
